pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "seraph"
version = "0.1.0"
description = "SERAPH Security Lake Dashboard - AWS Security Lake monitoring and query interface"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "SERAPH Team" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]
dependencies = [
    "boto3>=1.34.0",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "boto3-stubs[securitylake,athena,glue]>=1.34.0",
]
cdk = [
    "aws-cdk-lib>=2.120.0",
    "constructs>=10.3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["infrastructure/stacks"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
